+++ backend/core/src/main/java/org/sonarsource/sonarlint/core/tracking/XodusKnownFindingsStore.java
@@ -49,6 +49,7 @@ import org.sonarsource.sonarlint.core.serverconnection.storage.TarGzUtils;
 import org.sonarsource.sonarlint.core.serverconnection.storage.UuidBinding;
 
 import static java.util.Objects.requireNonNull;
+import static org.sonarsource.sonarlint.core.storage.XodusPurgeUtils.purgeOldTemporaryFiles;
 
 public class XodusKnownFindingsStore {
 
@@ -82,7 +83,7 @@ public class XodusKnownFindingsStore {
 
   public XodusKnownFindingsStore(Path backupDir, Path workDir) throws IOException {
     xodusDbDir = Files.createTempDirectory(workDir, KNOWN_FINDINGS_STORE);
-    purgeOldTemporaryFiles(workDir);
+    purgeOldTemporaryFiles(workDir, PURGE_NUMBER_OF_DAYS, KNOWN_FINDINGS_STORE + "*");
     backupFile = backupDir.resolve(BACKUP_TAR_GZ);
     if (Files.isRegularFile(backupFile)) {
       LOG.debug("Restoring previous known findings database from {}", backupFile);
@@ -98,23 +99,6 @@ public class XodusKnownFindingsStore {
       entityStore.registerCustomPropertyType(txn, Instant.class, new InstantBinding());
       entityStore.registerCustomPropertyType(txn, UUID.class, new UuidBinding());
     });
-  }
-
-  private static void purgeOldTemporaryFiles(Path workDir) {
-    if (Files.exists(workDir)) {
-      try (var stream = Files.newDirectoryStream(workDir, KNOWN_FINDINGS_STORE + "*")) {
-        for (var path : stream) {
-          var file = path.toFile();
-          var diff = new Date().getTime() - file.lastModified();
-          if (diff > PURGE_NUMBER_OF_DAYS * 24 * 60 * 60 * 1000) {
-            FileUtils.deleteQuietly(file);
-            LOG.debug("Successfully purged " + path);
-          }
-        }
-      } catch (Exception e) {
-        LOG.error("Unable to purge old temporary files for findings");
-      }
-    }
   }
 
   public List<KnownFinding> loadIssuesForFile(String configurationScopeId, Path filePath) {
